Extract file functions to util
